Remove UIC and RCC configuration variables
The idea behind these changes is to use Python to load PyQt5.uic module and then use the compileUi and compileRcc options from there. It should save the whole headache of compilation and such. Anyway, the purpose of the commandline tools is for the **user** to be able to compile things, not an automated script.